# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NFL Season-Long Pick'em League API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface UserUpdateInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  name?: string;
  /** @nullable */
  avatar?: string | null;
}

export interface UserLogin {
  /** @minLength 1 */
  name: string;
}

export interface Match {
  id: number;
  week: number;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  winner?: string | null;
  isCompleted: boolean;
  /** @nullable */
  pointSpread?: string | null;
  /** @nullable */
  injuryWeatherFlags?: string | null;
  /** @nullable */
  gameTime?: string | null;
}

export interface Pick {
  id: number;
  userId: number;
  matchId: number;
  selectedTeam: string;
  isLock: boolean;
  pointsEarned: number;
  match?: Match;
}

export interface SinglePickInput {
  matchId: number;
  selectedTeam: string;
  isLock: boolean;
}

export interface BulkPickInput {
  userId: number;
  picks: SinglePickInput[];
}

export type AutofillInputMode = typeof AutofillInputMode[keyof typeof AutofillInputMode];


export const AutofillInputMode = {
  favorites: 'favorites',
  home: 'home',
  random: 'random',
} as const;

export interface AutofillInput {
  userId: number;
  mode: AutofillInputMode;
}

export interface LeaderboardEntry {
  userId: number;
  name: string;
  totalPoints: number;
  rank: number;
  correctPicks: number;
  totalPicks: number;
  badges: string[];
}

export interface TrendEntry {
  userId: number;
  name: string;
  weeklyPoints: number[];
}

export interface WeeklyScore {
  userId: number;
  name: string;
  points: number;
}

export interface WeeklyExtremes {
  week: number;
  topUsers: WeeklyScore[];
  bottomUsers: WeeklyScore[];
}

export interface PickPopularity {
  matchId: number;
  week: number;
  homeTeam: string;
  awayTeam: string;
  homePickPct: number;
  awayPickPct: number;
  homePickCount: number;
  awayPickCount: number;
  homePickerNames: string[];
  awayPickerNames: string[];
}

export type SeasonStatusMode = typeof SeasonStatusMode[keyof typeof SeasonStatusMode];


export const SeasonStatusMode = {
  'pre-season': 'pre-season',
  'in-season': 'in-season',
} as const;

export interface SeasonStatus {
  mode: SeasonStatusMode;
  lastCompletedWeek: number;
  seasonLocked: boolean;
}

export interface SmackMessage {
  id: number;
  name: string;
  message: string;
  timestamp: string;
}

export interface SmackMessageInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 1
     * @maxLength 280
     */
  message: string;
}

export interface MatchResult {
  winner: string;
}

export interface WebhookInput {
  webhookUrl: string;
  message: string;
}

export interface WebhookResult {
  success: boolean;
  message?: string;
}

export type SeasonModeInputMode = typeof SeasonModeInputMode[keyof typeof SeasonModeInputMode];


export const SeasonModeInputMode = {
  'pre-season': 'pre-season',
  'in-season': 'in-season',
} as const;

export interface SeasonModeInput {
  mode: SeasonModeInputMode;
}

export type ListMatchesParams = {
week?: number;
};