# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Season-Long Pick'em League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Log in or create user by name (no password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Existing user logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "201":
          description: New user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile (name and/or avatar)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # ── Matches ────────────────────────────────────────────────────────────────
...
  /admin/match/{matchId}/result:
    patch:
      operationId: setMatchResult
      tags: [admin]
      summary: Set the winner of a match (commissioner only)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchResult"
      responses:
        "200":
          description: Match updated, scores recalculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /admin/webhook:
    post:
      operationId: sendWebhookNotification
      tags: [admin]
      summary: Send a Discord/Slack webhook reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "200":
          description: Webhook sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"
  /admin/season:
    patch:
      operationId: updateSeasonMode
      tags: [admin]
      summary: Toggle season mode (pre-season / in-season)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonModeInput"
      responses:
        "200":
          description: Season mode updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonStatus"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
        avatar:
          type: ["string", "null"]
    UserLogin:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    Match:
      type: object
      required: [id, week, homeTeam, awayTeam, isCompleted]
      properties:
        id:
          type: integer
        week:
          type: integer
        homeTeam:
          type: string
        awayTeam:
          type: string
        winner:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        pointSpread:
          type: ["string", "null"]
        injuryWeatherFlags:
          type: ["string", "null"]
        gameTime:
          type: ["string", "null"]
...
    SeasonStatus:
      type: object
      required: [mode, lastCompletedWeek, seasonLocked]
      properties:
        mode:
          type: string
          enum: [pre-season, in-season]
        lastCompletedWeek:
          type: integer
        seasonLocked:
          type: boolean
...
    MatchResult:
      type: object
      required: [winner]
      properties:
        winner:
          type: string
    WebhookInput:
      type: object
      required: [webhookUrl, message]
      properties:
        webhookUrl:
          type: string
        message:
          type: string
    WebhookResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    SeasonModeInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [pre-season, in-season]