# lib/api-zod/src/generated/types/userUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NFL Season-Long Pick'em League API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdateInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  name?: string;
  /** @nullable */
  avatar?: string | null;
}